backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
beanie>=1.11.0
python-dotenv>=0.19.0
pydantic>=1.8.2
